# optimized textwidth(::Char) for ASCII