--- a/base/strings/unicode.jl
+++ b/base/strings/unicode.jl
@@ -259,6 +259,13 @@ julia> textwidth('⛵')
 ```
 """
 function textwidth(c::AbstractChar)
+    ismalformed(c) && return 1
+    i = codepoint(c)
+    i < 0x80 && @inbounds return Int(_textwidth_ascii[i%Int+1]) # ASCII fast path
+    Int(ccall(:utf8proc_charwidth, Cint, (UInt32,), i))
+end
+
+function textwidth(c::Char)
     b = bswap(reinterpret(UInt32, c)) # from isascii(c)
     b < 0x80 && @inbounds return Int(_textwidth_ascii[b%Int+1]) # ASCII fast path
     ismalformed(c) && return 1