--- a/base/strings/unicode.jl
+++ b/base/strings/unicode.jl
@@ -240,9 +240,6 @@ end
 
 ############################################################################
 
-# precomputed table of ASCII text widths
-const _textwidth_ascii = ntuple(i -> ccall(:utf8proc_charwidth, Cint, (UInt32,), i-1), 128)
-
 ## character column width function ##
 """
     textwidth(c)
@@ -261,13 +258,13 @@ julia> textwidth('⛵')
 function textwidth(c::AbstractChar)
     ismalformed(c) && return 1
     i = codepoint(c)
-    i < 0x80 && @inbounds return Int(_textwidth_ascii[i%Int+1]) # ASCII fast path
+    i < 0x7f && return Int(i >= 0x20); # ASCII fast path
     Int(ccall(:utf8proc_charwidth, Cint, (UInt32,), i))
 end
 
 function textwidth(c::Char)
     b = bswap(reinterpret(UInt32, c)) # from isascii(c)
-    b < 0x80 && @inbounds return Int(_textwidth_ascii[b%Int+1]) # ASCII fast path
+    b < 0x7f && return Int(b >= 0x20); # ASCII fast path
     ismalformed(c) && return 1
     Int(ccall(:utf8proc_charwidth, Cint, (UInt32,), c))
 end