For the common case of ASCII characters, I found in https://github.com/JuliaLang/julia/pull/55351/files#r1705746669 that `textwidth` can be sped up by ~~60%~~ 2.7x or more by adding a fast-path ASCII ~~lookup table~~ hard-coded rule.

The slowdown for non-ASCII characters seems negligible (1% or less).  Here, `textwidth2` is the new function, and `textwidth` is the old function (Julia 1.10.4, Apple M1 Pro):
```jl
julia> @btime textwidth('x'); @btime textwidth2('x')
  3.167 ns (0 allocations: 0 bytes)
  1.166 ns (0 allocations: 0 bytes)
1

julia> @btime textwidth('α'); @btime textwidth2('α')
  3.208 ns (0 allocations: 0 bytes)
  3.166 ns (0 allocations: 0 bytes)
1
```