@stevengj 2024-08-06
@jakobnissen, good catch, a tuple table indeed seems to be faster:
```jl
julia> @btime textwidth('x'); @btime textwidth2('x'); @btime textwidth3('x');
  3.166 ns (0 allocations: 0 bytes)
  1.958 ns (0 allocations: 0 bytes)
  1.125 ns (0 allocations: 0 bytes)
```
(`textwidth` is the old version, `textwidth2` is the array version, and `textwidth3` is the tuple version)